backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
sentence-transformers
pinecone-client
transformers
torch           
scikit-learn
pandas
numpy
python-multipart
Pillow